Previously, if you called one of the read or write methods, the read/write could possibly start AND finish before the method even returned.  E.g.:
[asyncSocket readDataToLength:15 timeout:5 tag:10];
// Read has already finished, and invoked your delegate method!
state = READING; // Oops!

This isn't what people generally suspect, and it runs a bit counter to the asynchronous idea.
Thus this has been changed. Calling a read or write method will no longer finish prior to the method returning. The read/write is guaranteed to finish until at least the next invocation of the run loop.


git-svn-id: http://cocoaasyncsocket.googlecode.com/svn/trunk@23 903ec77f-374c-0410-a8be-254fceba1445